requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai 
chromadb
pypdf
python-dotenv